+++ src/Colony.ts
@@ -1,6 +1,7 @@
 import {assimilationLocked} from './assimilation/decorator';
 import {$} from './caching/GlobalCache';
 import {log} from './console/log';
+import {Roles} from './creepSetups/setups';
 import {StoreStructure} from './declarations/typeGuards';
 // import {DirectivePraise} from './directives/colony/praise';
 import {DirectiveExtract} from './directives/resource/extract';
@@ -25,6 +26,7 @@ import {WorkerOverlord} from './overlords/core/worker';
 import {RandomWalkerScoutOverlord} from './overlords/scouting/randomWalker';
 import {profile} from './profiler/decorator';
 import {Abathur} from './resources/Abathur';
+import {ALL_ZERO_ASSETS} from './resources/map_resources';
 import {bunkerLayout, getPosFromBunkerCoord} from './roomPlanner/layouts/bunker';
 import {RoomPlanner} from './roomPlanner/RoomPlanner';
 import {LOG_STATS_INTERVAL, Stats} from './stats/stats';
@@ -108,10 +110,9 @@ export class Colony {
 	extensions: StructureExtension[];					// |
 	storage: StructureStorage | undefined;				// |
 	links: StructureLink[];								// |
-	availableLinks: StructureLink[];
-	// claimedLinks: StructureLink[];						// | Links belonging to hive cluseters excluding mining groups
-	// dropoffLinks: StructureLink[]; 						// | Links not belonging to a hiveCluster, used as dropoff
+	availableLinks: StructureLink[];					// | links available to claim
 	terminal: StructureTerminal | undefined;			// |
+	factory: StructureFactory | undefined;				// |
 	towers: StructureTower[];							// |
 	labs: StructureLab[];								// |
 	powerSpawn: StructurePowerSpawn | undefined;		// |
@@ -262,6 +263,7 @@ export class Colony {
 		this.links = this.room.links;
 		this.availableLinks = _.clone(this.room.links);
 		this.terminal = this.room.terminal && this.room.terminal.isActive() ? this.room.terminal : undefined;
+		this.factory = this.room.factory && this.room.factory.isActive() ? this.room.factory : undefined;
 		this.towers = this.room.towers;
 		this.labs = _.sortBy(_.filter(this.room.labs, lab => lab.my && lab.isActive()),
 							 lab => 50 * lab.pos.y + lab.pos.x); // Labs are sorted in reading order of positions
@@ -284,7 +286,7 @@ export class Colony {
 		this.repairables = _.flatten(_.map(this.rooms, room => room.repairables));
 		this.rechargeables = _.flatten(_.map(this.rooms, room => room.rechargeables));
 		// Register assets
-		this.assets = this.getAllAssets();
+		this.assets = this.computeAssets();
 	}
 
 	/**
@@ -306,8 +308,8 @@ export class Colony {
 		$.set(this, 'spawns', () => _.sortBy(_.filter(this.room.spawns,
 													  spawn => spawn.my && spawn.isActive()), spawn => spawn.ref));
 		$.set(this, 'storage', () => this.room.storage && this.room.storage.isActive() ? this.room.storage : undefined);
-		// this.availableLinks = _.clone(this.room.links);
 		$.set(this, 'terminal', () => this.room.terminal && this.room.terminal.isActive() ? this.room.terminal : undefined);
+		$.set(this, 'factory', () => this.room.factory && this.room.factory.isActive() ? this.room.factory : undefined);
 		$.set(this, 'labs', () => _.sortBy(_.filter(this.room.labs, lab => lab.my && lab.isActive()),
 										   lab => 50 * lab.pos.y + lab.pos.x));
 		this.pos = (this.storage || this.terminal || this.spawns[0] || this.controller).pos;
@@ -333,7 +335,7 @@ export class Colony {
 		$.set(this, 'tombstones', () => _.flatten(_.map(this.rooms, room => room.tombstones)), 5);
 		this.drops = _.merge(_.map(this.rooms, room => room.drops));
 		// Register assets
-		this.assets = this.getAllAssets();
+		this.assets = this.computeAssets();
 	}
 
 	/**
@@ -341,13 +343,13 @@ export class Colony {
 	 */
 	private refreshRoomObjects(): void {
 		$.refresh(this, 'controller', 'extensions', 'links', 'towers', 'powerSpawn', 'nuker', 'observer', 'spawns',
-				  'storage', 'terminal', 'labs', 'sources', 'extractors', 'constructionSites', 'repairables',
+				  'storage', 'terminal', 'factory', 'labs', 'sources', 'extractors', 'constructionSites', 'repairables',
 				  'rechargeables');
 		$.set(this, 'constructionSites', () => _.flatten(_.map(this.rooms, room => room.constructionSites)), 10);
 		$.set(this, 'tombstones', () => _.flatten(_.map(this.rooms, room => room.tombstones)), 5);
 		this.drops = _.merge(_.map(this.rooms, room => room.drops));
 		// Re-compute assets
-		this.assets = this.getAllAssets();
+		this.assets = this.computeAssets();
 	}
 
 	/**
@@ -532,20 +534,17 @@ export class Colony {
 	}
 
 	/**
-	 * Summarizes the total of all resources in colony store structures, labs, and some creeps
+	 * Summarizes the total of all resources in colony store structures, labs, and some creeps. Will always return
+	 * 0 for an asset that it has none of (not undefined)
 	 */
-	private getAllAssets(verbose = false): { [resourceType: string]: number } {
-		// if (this.name == 'E8S45') verbose = true; // 18863
+	private computeAssets(verbose = false): { [resourceType: string]: number } {
 		// Include storage structures, lab contents, and manager carry
-		const stores = _.map(<StoreStructure[]>_.compact([this.storage, this.terminal]), s => s.store);
-		const creepCarriesToInclude = _.map(this.creeps, creep => creep.carry) as { [resourceType: string]: number }[];
-		const labContentsToInclude = _.map(_.filter(this.labs, lab => !!lab.mineralType), lab =>
-			({[<string>lab.mineralType]: lab.mineralAmount})) as { [resourceType: string]: number }[];
-		const allAssets: { [resourceType: string]: number } = mergeSum([
-																		   ...stores,
-																		   ...creepCarriesToInclude,
-																		   ...labContentsToInclude
-																	   ]);
+		const assetStructures = _.compact([this.storage, this.terminal, this.factory, ...this.labs]);
+		const assetCreeps = [...this.getCreepsByRole(Roles.queen), ...this.getCreepsByRole(Roles.manager)];
+		const assetStores = _.map([...assetStructures, ...assetCreeps], thing => thing!.store);
+
+		const allAssets = mergeSum([...assetStores, ALL_ZERO_ASSETS]);
+
 		if (verbose) log.debug(`${this.room.print} assets: ` + JSON.stringify(allAssets));
 		return allAssets;
 	}